Fix 11 bugs across ISP pipeline modules
- crop.py: .T.shape stored (W,H) instead of (H,W); -0 slice produced empty array
- auto_exposure.py: BT.601 blue weight 0.144→0.114; clip bound 256→255
- eotf.py: EOTF normalization used img.max() instead of bit-depth full-scale
- color_space_conversion.py: Y offset 2^(bit_depth/2)→2^(bit_depth-4) for non-8-bit
- scale.py: inconsistent // vs / in width downscale calculation
- hdr_merge.py: mid-grey constraint overwrote smoothness row in Debevec solver
- operators.py: highlight_rolloff knee mapped to 1.10 instead of 1.0
- temporal_nr.py: sequential in-place erosion caused directional bias in motion mask
- nearest_neighbor.py: missing index clamp risked out-of-bounds access
- non_local_means.py: output array was always uint8 regardless of input dtype
- dither.py: removed unused noise variable (dead code cleanup)

+++ modules/noise_reduction_2d/non_local_means.py
@@ -67,7 +67,7 @@ class NLM:
             input_image = np.round(255 * input_image).astype(np.uint8)
 
         # Declaring empty array for output image after denoising
-        denoised_out = np.empty(in_image.shape, dtype=np.uint8)
+        denoised_out = np.empty(in_image.shape, dtype=in_image.dtype)
 
         # Padding the input_image
         pads = window_size // 2
